refactor(aletheia): extract runtime::tool_adapters submodule (#2642)
Move CrossNousAdapter and SignalAdapter to a dedicated file. These are
trait-object adapters bridging tool services to the agora/nous layers.
